Al but vli test now compile on MSVC 9 and 10
git-svn-id: svn://localhost:7744/trunk/alps@5931 3665984a-f8b0-4bb5-b30c-cdcaecbe7ced

+++ applications/dmft/qmc/interaction_expansion2/selfenergy.cpp
@@ -398,7 +398,7 @@ void evaluate_selfenergy_measurement_itime_rs(const alps::results_type<HubbardIn
       Sz_err += Sz_i_err;
     }
     Sz_obs /= n_site;
-    Sz_err /= std::sqrt(n_site);
+    Sz_err /= std::sqrt(double(n_site));
     std::ofstream szstream("staggered_sz", std::ios::app);
     szstream << Sz_obs << "\t" << Sz_err << std::endl;
   }
